backend/requirements.txt
fastapi==0.115.12
motor==3.7.0
passlib==1.7.4
pydantic==2.11.3
pymongo==4.12.0
python-dotenv==1.1.0
python_jose==3.4.0
uvicorn==0.34.2
email-validator==2.2.0
